Move some more files
